remove MTV resource, Move Resource constants to class
