+++ ocp_resources/datavolume.py
@@ -1,4 +1,4 @@
-from ocp_resources.constants import (
+from ocp_resources.utils.constants import (
     TIMEOUT_1MINUTE,
     TIMEOUT_2MINUTES,
     TIMEOUT_4MINUTES,
