+++ ocp_resources/deployment.py
@@ -3,7 +3,7 @@
 from typing import Any, Dict, Optional
 from timeout_sampler import TimeoutSampler, TimeoutWatch
 
-from ocp_resources.constants import PROTOCOL_ERROR_EXCEPTION_DICT, TIMEOUT_4MINUTES
+from ocp_resources.utils.constants import PROTOCOL_ERROR_EXCEPTION_DICT, TIMEOUT_4MINUTES
 from ocp_resources.resource import NamespacedResource, MissingRequiredArgumentError
