+++ ocp_resources/endpoint_slice.py
@@ -1,4 +1,4 @@
-from ocp_resources.constants import TIMEOUT_4MINUTES
+from ocp_resources.utils.constants import TIMEOUT_4MINUTES
 from ocp_resources.resource import MissingRequiredArgumentError, NamespacedResource
