+++ ocp_resources/job.py
@@ -1,6 +1,6 @@
 import kubernetes
 
-from ocp_resources.constants import TIMEOUT_4MINUTES
+from ocp_resources.utils.constants import TIMEOUT_4MINUTES
 from ocp_resources.resource import NamespacedResource
