+++ ocp_resources/node_network_state.py
@@ -2,7 +2,7 @@ import time
 
 from kubernetes.dynamic.exceptions import ConflictError
 
-from ocp_resources.constants import TIMEOUT_4MINUTES
+from ocp_resources.utils.constants import TIMEOUT_4MINUTES
 from ocp_resources.resource import Resource
 from timeout_sampler import TimeoutSampler
