+++ ocp_resources/persistent_volume_claim.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from ocp_resources.constants import TIMEOUT_4MINUTES
+from ocp_resources.utils.constants import TIMEOUT_4MINUTES
 from ocp_resources.resource import NamespacedResource
