+++ ocp_resources/resource.py
@@ -29,7 +29,7 @@ from packaging.version import Version
 from simple_logger.logger import get_logger, logging
 from urllib3.exceptions import MaxRetryError
 
-from ocp_resources.constants import (
+from ocp_resources.utils.constants import (
     DEFAULT_CLUSTER_RETRY_EXCEPTIONS,
     NOT_FOUND_ERROR_EXCEPTION_DICT,
     PROTOCOL_ERROR_EXCEPTION_DICT,
@@ -46,7 +46,8 @@ from timeout_sampler import (
     TimeoutWatch,
 )
 from ocp_resources.exceptions import MissingRequiredArgumentError, MissingResourceResError
-from ocp_resources.utils import skip_existing_resource_creation_teardown
+from ocp_resources.utils.resource_constants import ResourceConstants
+from ocp_resources.utils.utils import skip_existing_resource_creation_teardown
 
 
 LOGGER = get_logger(name=__name__)
@@ -273,7 +274,7 @@ class ClassProperty:
         return self.func(owner)
 
 
-class Resource:
+class Resource(ResourceConstants):
     """
     Base class for API resources
     """
@@ -283,65 +284,65 @@ class Resource:
     singular_name: str = ""
     timeout_seconds: int = TIMEOUT_1MINUTE
 
-    class Status:
-        SUCCEEDED: str = "Succeeded"
-        FAILED: str = "Failed"
-        DELETING: str = "Deleting"
-        DEPLOYED: str = "Deployed"
-        PENDING: str = "Pending"
-        COMPLETED: str = "Completed"
-        RUNNING: str = "Running"
-        READY: str = "Ready"
-        TERMINATING: str = "Terminating"
-        ERROR: str = "Error"
-        COMPLETE: str = "Complete"
-        DEPLOYING: str = "Deploying"
-        SCHEDULING_DISABLED = "Ready,SchedulingDisabled"
-        CRASH_LOOPBACK_OFF = "CrashLoopBackOff"
-        IMAGE_PULL_BACK_OFF = "ImagePullBackOff"
-        ERR_IMAGE_PULL = "ErrImagePull"
-        ACTIVE = "Active"
-
-    class Condition:
-        UPGRADEABLE: str = "Upgradeable"
-        AVAILABLE: str = "Available"
-        DEGRADED: str = "Degraded"
-        PROGRESSING: str = "Progressing"
-        CREATED: str = "Created"
-        RECONCILE_COMPLETE: str = "ReconcileComplete"
-        READY: str = "Ready"
-        FAILING: str = "Failing"
-
-        class Status:
-            TRUE: str = "True"
-            FALSE: str = "False"
-            UNKNOWN: str = "Unknown"
-
-        class Phase:
-            INSTALL_READY: str = "InstallReady"
-            SUCCEEDED: str = "Succeeded"
-
-        class Reason:
-            ALL_REQUIREMENTS_MET: str = "AllRequirementsMet"
-            INSTALL_SUCCEEDED: str = "InstallSucceeded"
-            NETWORK_ATTACHMENT_DEFINITION_READY: str = "NetworkAttachmentDefinitionReady"
-            SYNC_ERROR: str = "SyncError"
-
-        class Type:
-            NETWORK_READY: str = "NetworkReady"
-            SUCCESSFUL: str = "Successful"
-            RUNNING: str = "Running"
-
-    class Type:
-        CLUSTER_IP = "ClusterIP"
-        NODE_PORT = "NodePort"
-        LOAD_BALANCER = "LoadBalancer"
-
-    class Interface:
-        class State:
-            UP: str = "up"
-            DOWN: str = "down"
-            ABSENT: str = "absent"
+    # class Status:
+    #     SUCCEEDED: str = "Succeeded"
+    #     FAILED: str = "Failed"
+    #     DELETING: str = "Deleting"
+    #     DEPLOYED: str = "Deployed"
+    #     PENDING: str = "Pending"
+    #     COMPLETED: str = "Completed"
+    #     RUNNING: str = "Running"
+    #     READY: str = "Ready"
+    #     TERMINATING: str = "Terminating"
+    #     ERROR: str = "Error"
+    #     COMPLETE: str = "Complete"
+    #     DEPLOYING: str = "Deploying"
+    #     SCHEDULING_DISABLED = "Ready,SchedulingDisabled"
+    #     CRASH_LOOPBACK_OFF = "CrashLoopBackOff"
+    #     IMAGE_PULL_BACK_OFF = "ImagePullBackOff"
+    #     ERR_IMAGE_PULL = "ErrImagePull"
+    #     ACTIVE = "Active"
+    #
+    # class Condition:
+    #     UPGRADEABLE: str = "Upgradeable"
+    #     AVAILABLE: str = "Available"
+    #     DEGRADED: str = "Degraded"
+    #     PROGRESSING: str = "Progressing"
+    #     CREATED: str = "Created"
+    #     RECONCILE_COMPLETE: str = "ReconcileComplete"
+    #     READY: str = "Ready"
+    #     FAILING: str = "Failing"
+    #
+    #     class Status:
+    #         TRUE: str = "True"
+    #         FALSE: str = "False"
+    #         UNKNOWN: str = "Unknown"
+    #
+    #     class Phase:
+    #         INSTALL_READY: str = "InstallReady"
+    #         SUCCEEDED: str = "Succeeded"
+    #
+    #     class Reason:
+    #         ALL_REQUIREMENTS_MET: str = "AllRequirementsMet"
+    #         INSTALL_SUCCEEDED: str = "InstallSucceeded"
+    #         NETWORK_ATTACHMENT_DEFINITION_READY: str = "NetworkAttachmentDefinitionReady"
+    #         SYNC_ERROR: str = "SyncError"
+    #
+    #     class Type:
+    #         NETWORK_READY: str = "NetworkReady"
+    #         SUCCESSFUL: str = "Successful"
+    #         RUNNING: str = "Running"
+    #
+    # class Type:
+    #     CLUSTER_IP = "ClusterIP"
+    #     NODE_PORT = "NodePort"
+    #     LOAD_BALANCER = "LoadBalancer"
+    #
+    # class Interface:
+    #     class State:
+    #         UP: str = "up"
+    #         DOWN: str = "down"
+    #         ABSENT: str = "absent"
 
     class ApiGroup:
         AAQ_KUBEVIRT_IO: str = "aaq.kubevirt.io"
