+++ ocp_resources/utils/constants.py
