+++ ocp_resources/utils/utils.py
